Add interfaces to translator
